cua-agent/requirements.txt
openai>=1.40.0
playwright==1.50.0
pillow==11.1.0
python-dotenv==1.0.1
requests==2.32.3
